requirements.txt
# Зависимости для проекта "Второй мозг"
# Персональный голосовой интеллект

# Telegram Bot API
python-telegram-bot==20.7

# Работа с переменными окружения
python-dotenv==1.0.0

# База данных SQLite (встроена в Python)
# psycopg2-binary==2.9.9  # Для PostgreSQL (если понадобится в будущем)

# Обработка аудио (если понадобится)
# pydub==0.25.1
# librosa==0.10.1

# Машинное обучение и NLP
openai==1.3.0  # DeepSeek совместим с OpenAI API
# langchain==0.0.350
# sentence-transformers==2.2.2

# Векторные базы данных (для будущих этапов)
# faiss-cpu==1.7.4
# qdrant-client==1.6.9
